Reliable Damage
Fixed health stuff with player and enemy

+++ GnomadGame/Assets/Entities/Enemies/Gorb/Scripts/GorbMovementStats.cs
@@ -22,6 +22,6 @@ public class GorbStats : EnemyMovementStats
     //Slashing Variabels
     public new const float attackDuration = 1.5f;
     public new const float attackCooldown = 2.5f;
-    public new const int attackDamage = 20;
+    public new const float attackDamage = 20f;
 
 }
